Remove npm package `bootstrap-select'
* The `bootstrap-select' project at https://github.com/snapappointments/bootstrap-select/ is dead
* Update Bootstrap from v4.5.0 to v4.6.2
* Replace with a regular <select> HTML element with bootstrap classes

`SiteScriptTagHelper` - changed behavior:
If `OnContentLoaded` is set, it checks for `document.readState != 'loading'`.
When not 'loading', the callback is invoked directly, otherwise the callback will be called on `document.DOMContentLoaded`

`RoundSelector` and `TeamEditor` components:
SiteScriptTagHelper.Inline is set to false. This goes together with the changed behavior of `SiteScriptTagHelper` (see above)